% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{mui.bgcolors}
\alias{mui.bgcolors}
\alias{mui.fgcolors}
\title{Metro UI colors}
\usage{
mui.bgcolors()

mui.fgcolors()
}
\value{
The vector of available colors.
}
\description{
Returns the vector of available colors.
}
