% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{multiActionButton}
\alias{multiActionButton}
\title{Multi action button}
\usage{
multiActionButton(
  rotate = TRUE,
  bg = NULL,
  fg = NULL,
  icon = NULL,
  direction = "right",
  subButtons
)
}
\arguments{
\item{rotate}{logical, whether to rotate the icon upon clicking the main
button}

\item{bg}{background color of the button}

\item{fg}{color of the icon}

\item{icon}{icon name}

\item{direction}{the opening direction, \code{"top"}, \code{"right"},
\code{"bottom"}, or \code{"left"}}

\item{subButtons}{a list of buttons created with \code{\link{subButton}}}
}
\description{
Creates a multi action button for usage in a Shiny UI.
}
\examples{
library(shiny)
library(multiActionButton)

ui <- fluidPage(

  tags$head(
    tags$style(HTML("pre {font-size: 18px; font-weight: bold;}"))
  ),

  br(),

  multiActionButton(
    rotate = TRUE, bg = "darkMagenta", fg = "white",
    icon = "plus", direction = "right",
    subButtons = list(
      subButton("user-plus", bg = "blue", icon = "user-plus", iconSize = "4x"),
      subButton("library", bg = "teal", icon = "library", iconSize = "4x"),
      subButton("alarm", bg = "pink", icon = "alarm", iconSize = "4x"),
      subButton("power", bg = "lime", icon = "power", iconSize = "4x"),
      subButton("lock", bg = "orange", icon = "lock", iconSize = "4x")
    )
  ),

  br(), br(), br(),

  tags$fieldset(
    tags$legend("Buttons states"),
    verbatimTextOutput("buttons")
  )

)

server <- function(input, output){

  output[["buttons"]] <- renderPrint({
    states <- c(
      paste0("user-plus: ", input[["user-plus"]]),
      paste0("library: ",   input[["library"]]),
      paste0("alarm: ",     input[["alarm"]]),
      paste0("power: ",     input[["power"]]),
      paste0("lock: ",      input[["lock"]])
    )
    cat(states, sep = "\n")
  })

}

if(interactive()){
  shinyApp(ui, server)
}
}
