% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{shinyMetroUIcolors}
\alias{shinyMetroUIcolors}
\title{Shiny app displaying the metro 4 background colors}
\usage{
shinyMetroUIcolors()
}
\description{
Launch a shiny app showing all available metro 4 background colors.
}
