% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{shinyMUIcolors}
\alias{shinyMUIcolors}
\title{Shiny app: metro UI background colors}
\usage{
shinyMUIcolors()
}
\value{
No returned value, the function just launches a Shiny app.
}
\description{
Shiny app showing all available metro UI background colors.
}
