% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{shinyMUIdemo}
\alias{shinyMUIdemo}
\title{Shiny app: multi action button}
\usage{
shinyMUIdemo()
}
\value{
No returned value, this function just runs a Shiny app.
}
\description{
Shiny app demonstrating the multi action button.
}
