% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{shinyMUIicons}
\alias{shinyMUIicons}
\title{Shiny app: metro UI icons}
\usage{
shinyMUIicons()
}
\value{
No returned value, the function just launches a Shiny app.
}
\description{
Launch a shiny app displaying all available metro UI icons.
}
