% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{subButton}
\alias{subButton}
\title{Sub-button}
\usage{
subButton(
  id,
  bg = NULL,
  fg = NULL,
  icon = NULL,
  tooltip = NULL,
  onclick = NULL
)
}
\arguments{
\item{id}{button id}

\item{bg}{background color of the button}

\item{fg}{color of the icon}

\item{icon}{icon name}

\item{tooltip}{character string to be shown as tooltip}

\item{onclick}{the \code{onclick} attribute of the button}
}
\value{
A \code{shiny.tag} object for usage in
  \code{\link{multiActionButton}}.
}
\description{
Creates a sub-button for usage in
  \code{\link{multiActionButton}}.
}
