% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{mui.bgcolors}
\alias{mui.bgcolors}
\alias{mui.fgcolors}
\title{Metro UI colors}
\usage{
mui.bgcolors()

mui.fgcolors()
}
\value{
A character vector, the names of the available background and
  foreground colors.
}
\description{
Available metro UI colors.
}
